category/games/2026/02/20260208-124720-000-vector-lemmings-path-bridge/pyproject.toml
[project]
name = "vector-lemmings-path-bridge"
version = "0.1.0"
description = "Guide autonomous explorers to the exit by assigning path-finding roles"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
